% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_var.R
\name{flatten_var}
\alias{flatten_var}
\title{Flattens a var-result to a list of data.frames}
\usage{
flatten_var(x, se = TRUE)
}
\arguments{
\item{x}{a result of \code{\link[vars]{VAR}}}

\item{se}{if the standard errors should be reported}
}
\value{
a list of data.frames containing the estimates \code{est} and the
p-values \code{p-value}. If \code{se = TRUE}, the standard errors are also
reported as \code{se}.
}
\description{
Flattens a var-result to a list of data.frames
}
\examples{
library(vars)
data(Canada)
res <- VAR(Canada, p = 2, type = "none")

summary(res)
flatten_var(res)
flatten_var(res, se = FALSE)
}
