% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfy.R
\name{ggfy}
\alias{ggfy}
\title{Plots a VAR}
\usage{
ggfy(x, single_plot = TRUE, plot = TRUE)
}
\arguments{
\item{x}{an IRF-result as returned from \code{\link[vars]{irf}}}

\item{single_plot}{if all facets should be plotted in a single graph,
if FALSE, a list of plots for each from-variable is returned}

\item{plot}{if FALSE only the plotting data is returned instead of actual plots}
}
\value{
a data_frame if \code{plot == FALSE},
a single ggplot if \code{plot == TRUE, single_plot == TRUE} or
a list of ggplots if \code{plot == TRUE, single_plot == FALSE}
}
\description{
Plots a VAR
}
\examples{
library(vars)
data(Canada)

varres <- VAR(Canada, p = 2, type = "none")
# plot the VAR result
ggfy(varres)

# to extract the data only
ggfy(varres, plot = FALSE)

# to get a list of plots
plot_list <- ggfy(varres, single_plot = FALSE)

# IRF
irfres <- irf(varres)
ggfy(irfres)

# to extract the data only
ggfy(irfres, plot = FALSE)

# to get a list of plots
plot_list <- ggfy(irfres, single_plot = FALSE)

}
