% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulize.R
\name{tabulize}
\alias{tabulize}
\alias{tabulise}
\title{Creates a LaTeX or HTML table for a VAR result}
\usage{
tabulize(x, format = "latex", digits = 3, p = 0.1, se_values = TRUE,
  stars = c(`***` = 0, `**` = 0.001, `*` = 0.01, . = 0.05, ` ` = 0.1))

tabulise(x, format = "latex", digits = 3, p = 0.1, se_values = TRUE,
  stars = c(`***` = 0, `**` = 0.001, `*` = 0.01, . = 0.05, ` ` = 0.1))
}
\arguments{
\item{x}{a varest restult, as returned from \code{\link[vars]{VAR}}}

\item{format}{either "html" or "latex" depending on the desired output format}

\item{digits}{number of digits to display}

\item{p}{p-values threshold below which the estimate is displayed as bold,
if the estimate should not be boldened, use \code{p = 0}.}

\item{se_values}{if TRUE, the standard errors are reported instead of the
p-values, defaults to TRUE}

\item{stars}{a named vector of the signs, if set to \code{NULL} no stars are
shown}
}
\value{
latex or html string depending on the specified format
}
\description{
Creates a LaTeX or HTML table for a VAR result
}
\examples{
library(vars)
data(Canada)
varres <- VAR(Canada, p = 2, type = "none")

# latex table
tabulize(varres)

# html table
tabulize(varres, "html")

# if you want to display 5 digits
tabulize(varres, digits = 5)

# if you want to display 2 digits and a p-value threshold of 0.05
tabulize(varres, digits = 2, p = 0.05)

# disable bold
tabulize(varres, p = 0)

# show p-values instead of standard errors in the second row per variable
tabulize(varres, se_values = FALSE)

# dont show stars
tabulize(varres, stars = NULL)

# provide custome stars
stars_new <- c("***" = 0, "**" = 0.01, "*" = 0.05, " " = 0.1)
tabulize(varres, stars = stars_new, se_values = FALSE)
}
